fix issue with overflowing int64
