+++ adapters/repos/db/batch_reference_integration_test.go
@@ -22,12 +22,12 @@ import (
 	"time"
 
 	"github.com/go-openapi/strfmt"
-	"github.com/semi-technologies/weaviate/adapters/repos/db/vector/hnsw"
 	"github.com/semi-technologies/weaviate/entities/additional"
 	"github.com/semi-technologies/weaviate/entities/filters"
 	"github.com/semi-technologies/weaviate/entities/models"
 	"github.com/semi-technologies/weaviate/entities/schema"
 	"github.com/semi-technologies/weaviate/entities/schema/crossref"
+	enthnsw "github.com/semi-technologies/weaviate/entities/vectorindex/hnsw"
 	"github.com/semi-technologies/weaviate/usecases/config"
 	"github.com/semi-technologies/weaviate/usecases/objects"
 	"github.com/semi-technologies/weaviate/usecases/traverser"
@@ -63,7 +63,7 @@ func Test_AddingReferencesInBatches(t *testing.T) {
 		Objects: &models.Schema{
 			Classes: []*models.Class{
 				{
-					VectorIndexConfig:   hnsw.NewDefaultUserConfig(),
+					VectorIndexConfig:   enthnsw.NewDefaultUserConfig(),
 					InvertedIndexConfig: invertedConfig(),
 					Class:               "AddingBatchReferencesTestTarget",
 					Properties: []*models.Property{
@@ -75,7 +75,7 @@ func Test_AddingReferencesInBatches(t *testing.T) {
 					},
 				},
 				{
-					VectorIndexConfig:   hnsw.NewDefaultUserConfig(),
+					VectorIndexConfig:   enthnsw.NewDefaultUserConfig(),
 					InvertedIndexConfig: invertedConfig(),
 					Class:               "AddingBatchReferencesTestSource",
 					Properties: []*models.Property{
