+++ adapters/repos/db/multi_shard_integration_test.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/go-openapi/strfmt"
 	"github.com/google/uuid"
-	"github.com/semi-technologies/weaviate/adapters/repos/db/vector/hnsw"
 	"github.com/semi-technologies/weaviate/adapters/repos/db/vector/hnsw/distancer"
 	"github.com/semi-technologies/weaviate/entities/additional"
 	"github.com/semi-technologies/weaviate/entities/filters"
@@ -34,6 +33,7 @@ import (
 	"github.com/semi-technologies/weaviate/entities/schema/crossref"
 	"github.com/semi-technologies/weaviate/entities/search"
 	"github.com/semi-technologies/weaviate/entities/searchparams"
+	enthnsw "github.com/semi-technologies/weaviate/entities/vectorindex/hnsw"
 	"github.com/semi-technologies/weaviate/usecases/config"
 	"github.com/semi-technologies/weaviate/usecases/objects"
 	"github.com/semi-technologies/weaviate/usecases/sharding"
@@ -160,7 +160,7 @@ func Test_MultiShardJourneys_BM25_Search(t *testing.T) {
 	t.Run("prepare", func(t *testing.T) {
 		class := &models.Class{
 			Class:             className,
-			VectorIndexConfig: hnsw.NewDefaultUserConfig(),
+			VectorIndexConfig: enthnsw.NewDefaultUserConfig(),
 			InvertedIndexConfig: &models.InvertedIndexConfig{
 				CleanupIntervalSeconds: 60,
 			},
@@ -809,7 +809,7 @@ func bruteForceObjectsByQuery(objs []*models.Object,
 func testClassesForImporting() []*models.Class {
 	return []*models.Class{
 		{
-			VectorIndexConfig:   hnsw.NewDefaultUserConfig(),
+			VectorIndexConfig:   enthnsw.NewDefaultUserConfig(),
 			InvertedIndexConfig: invertedConfig(),
 			Class:               "TestClass",
 			Properties: []*models.Property{
@@ -836,7 +836,7 @@ func testClassesForImporting() []*models.Class {
 			},
 		},
 		{
-			VectorIndexConfig:   hnsw.NewDefaultUserConfig(),
+			VectorIndexConfig:   enthnsw.NewDefaultUserConfig(),
 			InvertedIndexConfig: invertedConfig(),
 			Class:               "TestRefClass",
 			Properties: []*models.Property{
