+++ adapters/repos/db/vector/hnsw/config.go
@@ -12,14 +12,8 @@
 package hnsw
 
 import (
-	"encoding/json"
-	"fmt"
-
-	"github.com/pkg/errors"
 	"github.com/semi-technologies/weaviate/adapters/repos/db/vector/hnsw/distancer"
 	"github.com/semi-technologies/weaviate/entities/errorcompounder"
-	"github.com/semi-technologies/weaviate/entities/schema"
-	ent "github.com/semi-technologies/weaviate/entities/vectorindex/hnsw"
 	"github.com/semi-technologies/weaviate/usecases/monitoring"
 	"github.com/sirupsen/logrus"
 )
@@ -66,155 +60,4 @@ func (c Config) Validate() error {
 	}
 
 	return ec.ToError()
-}
-
-// ParseUserConfig from an unknown input value, as this is not further
-// specified in the API to allow of exchanging the index type
-func ParseUserConfig(input interface{}) (schema.VectorIndexConfig, error) {
-	uc := ent.UserConfig{}
-	uc.SetDefaults()
-
-	if input == nil {
-		return uc, nil
-	}
-
-	asMap, ok := input.(map[string]interface{})
-	if !ok || asMap == nil {
-		return uc, fmt.Errorf("input must be a non-nil map")
-	}
-
-	if err := optionalIntFromMap(asMap, "maxConnections", func(v int) {
-		uc.MaxConnections = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "cleanupIntervalSeconds", func(v int) {
-		uc.CleanupIntervalSeconds = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "efConstruction", func(v int) {
-		uc.EFConstruction = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "ef", func(v int) {
-		uc.EF = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "dynamicEfFactor", func(v int) {
-		uc.DynamicEFFactor = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "dynamicEfMax", func(v int) {
-		uc.DynamicEFMax = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "dynamicEfMin", func(v int) {
-		uc.DynamicEFMin = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "vectorCacheMaxObjects", func(v int) {
-		uc.VectorCacheMaxObjects = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalIntFromMap(asMap, "flatSearchCutoff", func(v int) {
-		uc.FlatSearchCutoff = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalBoolFromMap(asMap, "skip", func(v bool) {
-		uc.Skip = v
-	}); err != nil {
-		return uc, err
-	}
-
-	if err := optionalStringFromMap(asMap, "distance", func(v string) {
-		uc.Distance = v
-	}); err != nil {
-		return uc, err
-	}
-
-	return uc, nil
-}
-
-func optionalIntFromMap(in map[string]interface{}, name string,
-	setFn func(v int),
-) error {
-	value, ok := in[name]
-	if !ok {
-		return nil
-	}
-
-	var asInt64 int64
-	var err error
-
-	// depending on whether we get the results from disk or from the REST API,
-	// numbers may be represented slightly differently
-	switch typed := value.(type) {
-	case json.Number:
-		asInt64, err = typed.Int64()
-	case float64:
-		asInt64 = int64(typed)
-	}
-	if err != nil {
-		return errors.Wrapf(err, "%s", name)
-	}
-
-	setFn(int(asInt64))
-	return nil
-}
-
-func optionalBoolFromMap(in map[string]interface{}, name string,
-	setFn func(v bool),
-) error {
-	value, ok := in[name]
-	if !ok {
-		return nil
-	}
-
-	asBool, ok := value.(bool)
-	if !ok {
-		return nil
-	}
-
-	setFn(asBool)
-	return nil
-}
-
-func optionalStringFromMap(in map[string]interface{}, name string,
-	setFn func(v string),
-) error {
-	value, ok := in[name]
-	if !ok {
-		return nil
-	}
-
-	asString, ok := value.(string)
-	if !ok {
-		return nil
-	}
-
-	setFn(asString)
-	return nil
-}
-
-func NewDefaultUserConfig() ent.UserConfig {
-	uc := ent.UserConfig{}
-	uc.SetDefaults()
-	return uc
 }
